requirements.txt
langchain
langchain-community
langchain-text-splitters
sentence-transformers
faiss-cpu
transformers
torch
pydantic
ollama
numpy
